% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_check.R
\name{argument_check}
\alias{argument_check}
\title{argument_check}
\usage{
argument_check(x, name, type, len_check = FALSE, len_req = 1)
}
\arguments{
\item{x}{an argument.}

\item{name}{the name of the argument.}

\item{type}{a string indicating what type the argument should be.}

\item{len_check}{a logical scalar indicating whether the argument should be checked for length.}

\item{len_req}{a numeric scalar indicating the required length for the argument.}
}
\description{
A function for streamlining the argument-checking process. It is not exported.
}
