% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.R
\name{bold}
\alias{bold}
\title{bold}
\usage{
bold(x)
}
\arguments{
\item{x}{a character scalar or vector to be formatted.}
}
\description{
A function for easily formatting strings as bold or italicized using markdown or LaTeX syntax. `bold(x)` is a shortcut for `text_format(x, format = "bold", latex = FALSE)` and `italic(x)` is a shortcut for `text_format(x, format = "italic", latex = FALSE)`. `bold_tex(x)` is a shortcut for `text_format(x, format = "bold", latex = TRUE)` and `italic_tex(x)` is a shortcut for `text_format(x, format = "italic", latex = TRUE)`.
}
