% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R
\name{build_models}
\alias{build_models}
\title{build_models}
\usage{
build_models(outcome, predictors)
}
\arguments{
\item{outcome}{a character string representing the outcome variable.}

\item{predictors}{a list of one or more strings representing the predictors. If a vector is included in the list, the vector values are added to the model simultaneously.}
}
\description{
Creates a composite column using only columns in a data frame that have names that match a pattern.
}
