% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capply.R
\name{capply}
\alias{capply}
\title{capply}
\usage{
capply(.data, .f)
}
\arguments{
\item{.data}{a data frame.}

\item{.f}{a function to be applied to every cell of the data frame.}
}
\description{
A wrapper of apply that allows the user to quickly apply a function to every cell of a data frame.
}
