% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{center}
\usage{
center(x)
}
\arguments{
\item{x}{a numeric vector or similar.}
}
\description{
Centers a numeric vector. It is identical to running `scale(x, center = TRUE, scale = FALSE)`.
}
