% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centre.R
\name{centre}
\alias{centre}
\title{centre}
\usage{
centre(x)
}
\arguments{
\item{x}{a numeric vector or similar.}
}
\description{
Centers a numeric vector. It is identical to running `scale(x, center = TRUE, scale = FALSE)`.
}
