% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_check.R
\name{choice_check}
\alias{choice_check}
\title{choice_check}
\usage{
choice_check(x, name, choices)
}
\arguments{
\item{x}{an argument.}

\item{name}{the name of the argument}

\item{choices}{a vector of acceptable choices. The first string is used as the default.}
}
\description{
A function for streamlining the argument-checking process when multiple strings are accepted.
}
