% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{collapse}
\usage{
collapse(vector, sep = " + ")
}
\arguments{
\item{vector}{a vector of character strings. (e.g., `c("item 1")`)}

\item{sep}{a character string to separate the results. Defaults to " + ".}
}
\description{
A shortcut of `paste(x, collapse = y)` for collapsing character vectors into a single string separated by some character. Running `collapse(c("item 1", "item 2"), sep = " + ")` is identical to running `paste(c("item 1", "item 2"), collapse = " + ")`.
}
