% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_alpha.R
\name{column_alpha}
\alias{column_alpha}
\title{column_alpha}
\usage{
column_alpha(pattern, data, full = FALSE, verbose = FALSE,
  na.rm = TRUE)
}
\arguments{
\item{pattern}{a character string that is used to find the columns of interest. It can be a regular expression.}

\item{data}{a data frame.}

\item{full}{if TRUE, the full reliability analysis produced by the `psych` package is returned. If FALSE, only the raw alpha value is returned.}

\item{verbose}{specifies whether all column names should be listed, regardless of length.}
}
\description{
Estimates Cronbach's Alpha--an indicator of internal consistency--using only columns that have names that match a pattern. The analysis relies relies on `psych::alpha`.
}
\examples{
x <- data.frame(my_column_x  = 1:10,
                my_column_x2 = 1:10 * rnorm(10, mean = 1),
                my_column_x3 = 1:10 * rnorm(10, mean = 1),
                my_column_y  = 11:20,
                my_column_y2 = 11:20 * rnorm(10, mean = 1),
                my_column_y3 = 11:20 * rnorm(10, mean = 1))

column_alpha(pattern = "x", data = x, full = TRUE)
column_alpha(pattern = "x", data = x, full = FALSE)

}
