% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_message.R
\name{column_message}
\alias{column_message}
\title{column message}
\usage{
column_message(data, function_operation, verbose = FALSE)
}
\arguments{
\item{data}{the results of column_find.}

\item{function_operation}{the operation to be performed on the results of column_find.}

\item{verbose}{specifies whether all column names should be listed, regardless of length.}
}
\description{
A function for creating messages when an action is performed on the results of column_find.
}
