% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reliability.R
\name{column_reliability}
\alias{column_reliability}
\title{column_reliability}
\usage{
column_reliability(
  data,
  pattern = "",
  return = "all",
  spround = FALSE,
  message = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a data frame.}

\item{pattern}{a character string that is used to find the columns of interest. It can be a regular expression. Defaults to all columns in data.}

\item{return}{a string indicating whether the function should return Cronbach's alpha (`"alpha"`), McDonald's omega (`"omega"`), the average inter-item correlation (`"rij"`), or all three (`"all"`). Defaults to `"all"`.}

\item{spround}{a logical value indicating whether values should be rounded for printing. Defaults to FALSE.}

\item{message}{if TRUE, messages are generated telling the user which columns were used to calculate the estimates.}

\item{verbose}{specifies whether all column names used should be listed in the message, regardless of length.}
}
\description{
Estimates Cronbach's alpha, McDonald's omega, and the average inter-item correlation using only columns that have names that match a provided pattern. The analysis relies on `MBESS::ci.reliability`.
}
