% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rs.R
\name{compare_rs}
\alias{compare_rs}
\title{compare_rs}
\usage{
compare_rs(
  cor_mat,
  predictors,
  outcome,
  test = "hittner2003",
  adjust.p = "none",
  threshold = 0.05
)
}
\arguments{
\item{cor_mat}{a correlation matrix object created using the `corr.test()` function from the `{psych}` package.}

\item{predictors}{a vector of strings indicating the names of the predictors.}

\item{outcome}{a string indicating the name of the outcome variable.}

\item{test}{a string indicating the type of test to use to compare the correlations. Defaults to hittner2003.}

\item{adjust.p}{a string indicating what type of correction for multiple comparisons should be used. Defaults to "none."}

\item{threshold}{a numeric scalar indicating the p-value that should be used to determine if two correlations are signficantly different from eachother. Defaults to .05.}
}
\description{
Compares two correlations. Currently, `compare_rs()` only supports correlations that have one variable in common and that were calculated in the same sample.
}
