% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{d}
\usage{
d(x, shorten = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{shorten}{a logical value indicating whether only the n, mean, standard deviation, minimum, and maximum should be returned. Defaults to `TRUE`.}

\item{...}{additional arguments passed to `describe()`.}
}
\description{
A very simple wrapper for the `describe()` function from the `{psych}` package. In addition to having a shorter function name, it also returns fewer statistics by default.
}
