% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark.R
\name{dark}
\alias{dark}
\alias{dark_triad}
\alias{dark_tetrad}
\title{dark}
\usage{
dark(
  constellation = "triad",
  shorten = FALSE,
  shorten_length = 4,
  format = "none",
  ...
)

dark_triad(x, ...)

dark_tetrad(x, ...)
}
\arguments{
\item{constellation}{a character scalar specifying whether to return the names of the Dark Triad traits or the names of the Dark Tetrad traits. Defaults to "Triad".}

\item{shorten}{a logical scalar indicating whether to shorten the trait names. Defaults to `FALSE`.}

\item{shorten_length}{a numeric scaler indicating the number of letters to return (e.g., `4` would return "Mach" instead of "Machiavellianism"). Defaults to `4`.}

\item{format}{a character scalar indicating whether to italicize or bold the string. Defaults to "none.}

\item{...}{additional arguments are assed to `text_format`.}
}
\description{
A function for returning a vector of aversive personality trait names (e.g., the Dark Triad or Dark Tetrad traits).
}
