% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_aic.R
\name{delta_aic}
\alias{delta_aic}
\title{delta_aic}
\usage{
delta_aic(models)
}
\arguments{
\item{models}{a list of two or more linear models. Differences will be calculated in the order models are listed.}
}
\description{
Calculates changes to AIC across two or more models.
}
