% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_bic.R
\name{delta_bic}
\alias{delta_bic}
\title{delta_bic}
\usage{
delta_bic(models)
}
\arguments{
\item{models}{a list of linear models. Differences will be calculated in the order models are listed.}
}
\description{
Calculates changes to BIC across two or more models.
}
