% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_rsq.R
\name{delta_rsq}
\alias{delta_rsq}
\title{delta_rsq}
\usage{
delta_rsq(models, adjusted = FALSE)
}
\arguments{
\item{models}{a list of two or more linear models. Differences will be calculated in the order models are listed.}

\item{adjusted}{if FALSE (the default), non-adjusted R-Squared values are used. If TRUE, adjusted R-Squared values are used.}
}
\description{
Calculates change to R-Squared across two or more models.
}
