\name{dgidbPan}
\alias{dgidbPan}
\docType{data}
\title{
Druggable Genome Interaction Database for Cancer
}
\description{
  Druggable Genome Interaction Database annotation for Cancer Drugs and
  their targets
}
\usage{data("dgidbPan")}
\format{
 %% A data frame with 6 subjects.  One non-responder and 5 matched responders
  \describe{
    \item{\code{Gene}}{Gene Symbol}
    \item{\code{Drugs}}{All cancer drugs that target the gene}
 }
}
\details{
  
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\keyword{datasets}
