% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_diff.R
\name{first_diff}
\alias{first_diff}
\title{first_diff}
\usage{
first_diff(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Takes a vector and calculates the difference between the first element and the sum of the other elements.
}
