% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_distance_badge.R
\name{fitbit_distance_badge}
\alias{fitbit_distance_badge}
\title{fitbit_distance_badge}
\usage{
fitbit_distance_badge(x, metric = TRUE, fitbit_only = FALSE)
}
\arguments{
\item{x}{a numeric scalar representing a distance in kilometres or miles.}

\item{metric}{a logical scalar indicating whether `x` is a distance in kilometres (`metric = TRUE`) or miles (`metric = FALSE`). Defaults to `TRUE` (i.e., kilometres).}

\item{fitbit_only}{a logical scalar indicating whether only badges from Fitbit should be returned. Defaults to `FALSE`.}
}
\description{
Find the badge corresponding to a distance in miles or kilometres
}
\examples{
fitbit_distance_badge(1000)
}
