% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_distance_badges.R
\docType{data}
\name{fitbit_distance_badges}
\alias{fitbit_distance_badges}
\title{Fitbit Distance Badges}
\format{
A data frame with 37 rows and 4 columns:
\describe{
  \item{badge}{The name of the lifetime distance badge}
  \item{fitbit}{Whether or not the badge is from Fitbit}
  \item{distances_mi}{The badge distance in miles}
  \item{distances_km}{The badge distance in kilometres}
}
}
\usage{
fitbit_distance_badges
}
\description{
A dataset including the Fitbit lifetime distance badges. It also includes some additional badges that are not from Fitbit.
}
\keyword{datasets}
