\name{gcinfoPan}
\alias{gcinfoPan}
\docType{data}
\title{
Gene information data for 7574 genes
}
\description{
  Gene information Data for 7574 genes
}
\usage{data("gcinfoPan")}
\format{
  A data frame with gene information
  \describe{
    \item{\code{CHROM}}{Chromosome}
    \item{\code{START}}{Gene start position}
    \item{\code{STOP}}{Gene stop position}
    \item{\code{Gene.Sybmol}}{Gene}
    \item{\code{CodingLength}}{Gene coding length}
     }
}
\details{

}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}

\examples{
data(gcinfoPan)
str(gcinfoPan)
}
\keyword{datasets}
