\name{gcPanCO}
\alias{gcPanCO}
\docType{data}
\title{
Dataset with cqn-normalized RNA gene count data for TCGA colon cancer
}
\description{
  Dataset with cqn-normalized RNA gene count data for TCGA colon cancer
}
\usage{data("gcPanCO")}
\format{
  A data frame with genes (rows) and patients (columns).  
  \describe{
    \item{\code{TCGA-DM-A0XD}}{normalized gene count data for patient TCGA-DM-A0XD}
 }
}
\details{

}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}

\examples{
data(gcPanCO)
str(gcPanCO)
}
\keyword{datasets}
