\name{gcPanTNBC}
\alias{gcPanTNBC}
\docType{data}
\title{
Dataset with cqn-normalized RNA gene count data for TCGA samples
}
\description{
  Dataset with cqn-normalized RNA gene count data for TCGA samples
}
\usage{data("gcPanTNBC")}
\format{
  A data frame with 5490 genes (rows) and  79 patients (columns). 

}
\details{

}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}

\examples{
data(gcPanTNBC)
str(gcPanTNBC)
}
\keyword{datasets}
