\name{genelistPan}
\alias{genelistPan}
\docType{data}
\title{
Curated Cancer Gene list for panoply
}
\description{
  Curated Cancer Gene list for panoply
}
\usage{data("genelistPan")}
\format{
  A data frame with gene information
}
\details{

}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}

\examples{
data(genelistPan)
str(genelistPan)
}
\keyword{datasets}
