% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paste.R
\name{get_paste}
\alias{get_paste}
\title{get_paste}
\usage{
get_paste(...)
}
\arguments{
\item{...}{a list of strings to be combined and gotten.}
}
\description{
A function for streamlining get(paste()). It is not exported.
}
