% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_compare.R
\name{group_compare}
\alias{group_compare}
\title{group_compare}
\usage{
group_compare(data, cols, split, var.equal = FALSE, adjust.p = "none",
  adjust.d = FALSE, spround = FALSE, spround.p = TRUE,
  collapse = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{cols}{a vector of strings specifying what columns to compare.}

\item{split}{a string indicating the column that includes the grouping variable.}

\item{var.equal}{a logical value indicating whether equal variances should be assumed.}

\item{adjust.p}{a string indicating what type of correction for multiple comparisons should be used. Defaults to "none."}

\item{adjust.d}{if TRUE, adjusts for a small sample and returns Hedges' g intead of Cohen's D.}

\item{spround}{a logical value indicating whether values should be rounded for printing.}

\item{spround.p}{a logical value indicating whether, if spround = TRUE, p-values should also be sprounded. Defaults to TRUE.}

\item{collapse}{a logical value indicating whether means and SDs should be combined using paste_paren.}

\item{na.rm}{a logical value indicating whether `NA` values should be removed prior to computation.}
}
\description{
Creates a group comparison table. It (1) calculates an overall mean and sd, as well as a mean and sd for each group, (2) runs a two-samples t-test comparing both groups, and (3) calculates Cohen's d.
}
