% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenique.R
\name{lenique}
\alias{lenique}
\title{lenique}
\usage{
lenique(x)
}
\arguments{
\item{x}{a scalar, atomic vector, or list.}
}
\description{
A very simple wrapper that calculates the length of unique values in a vector in one step. It is identical to running `length(unique(x))`.`
}
