% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_merge.R
\name{mat_merge}
\alias{mat_merge}
\title{mat_merge}
\usage{
mat_merge(x_mat, y_mat, x_from = "lower", y_from = "lower",
  x_to = "lower", y_to = "upper", diagonal = 0)
}
\arguments{
\item{x_mat}{a matrix.}

\item{y_mat}{a second matrix.}

\item{x_from}{if "lower" values are drawn from below the diagonal of x_mat. If "upper" values are drawn from above the diagonal of x_mat.}

\item{y_from}{if "lower" values are drawn from below the diagonal of y_mat. If "upper" values are drawn from above the diagonal of y_mat.}

\item{x_to}{if "lower" values from x_mat are placed below the diagonal. If "upper" values from x_mat are placed above the diagonal.}

\item{y_to}{if "lower" values from y_mat are placed below the diagonal. If "upper" values from y_mat are placed above the diagonal.}

\item{diagonal}{the value used to fill the diagonal. Defaults to 0.}
}
\description{
Combines two matrices by drawing values from either below or above the diagonal and placing them below and above the diagonal.
}
