% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_ci.R
\name{paste_ci}
\alias{paste_ci}
\title{paste_ci}
\usage{
paste_ci(x, y)
}
\arguments{
\item{x}{a scalar or an atomic vector.}

\item{y}{a second scalar or atomic vector that will be wrapped in parentheses.}
}
\description{
Combines two numbers (e.g., .20 and .33) by wrapping them in square brackets (e.g., [.20, .33]). This function was made to streamline the creation of confidence interval values for outputting.
}
