% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_paren.R
\name{paste_paren}
\alias{paste_paren}
\title{paste_paren}
\usage{
paste_paren(x, y)
}
\arguments{
\item{x}{a scalar or an atomic vector.}

\item{y}{a second scalar or atomic vector that will be wrapped in parentheses.}
}
\description{
If only `x` is provided, it wraps the value in parentheses (e.g., 10.12 becomes "10.12"). If `x` and `y` are both provided, it combines the two numbers (e.g., 10.12 and 2.22) by wrapping the latter in parentheses (e.g., 10.12(2.22)). This part of the function was made to streamline the creation of tables that include cells formatted mean(sd).
}
