% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perble.R
\name{perble}
\alias{perble}
\title{perble}
\usage{
perble(x, tidy = TRUE)
}
\arguments{
\item{x}{a scalar, atomic vector, or list.}

\item{tidy}{by default, results are tidied into long data frame. Each group has its own row and there are four rows: group, count, proportion, and percent. If `tidy = FALSE`, a matrix is returned with count, proportion, and percent saved as row names. The group names are saved as column names.}
}
\description{
Extends `table` by including proportions and percentages.
}
