% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{profile}
\usage{
profile(data, method = "de")
}
\arguments{
\item{data}{a dataset containing the two columns you wish to conduct a profile similarity analysis on.}

\item{method}{a string indicating the type of profile similarity analysis you wish to conduct. Defaults to "de" or double-entry q correlations.}

\item{...}{optional arguments to be passed to cor.}
}
\description{
Calculates the profile similarity between two profiles. Currently only supports "de" (double-entry *q* correlations) and "r" (Pearson correlation) methods.
}
