% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{reverse}
\usage{
reverse(x, low = 1, high = 5)
}
\arguments{
\item{x}{a numeric scalar to be reverse-scored.}

\item{low}{a numeric scalar representing the lowest possible value. Defaults to `1`.}

\item{high}{a numeric scalar representing the highest possible value. Defaults to `5`.}
}
\description{
Reverse scores a value based on (1) the value (i.e., `x`), (2) the lowest possible value (i.e., `low`), and (3) the highest possible value (i.e., `high`).
}
