% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_qualtrics.R
\docType{data}
\name{states_qualtrics}
\alias{states_qualtrics}
\title{states_qualtrics}
\format{
An object of class \code{data.frame} with 53 rows and 2 columns.
}
\usage{
states_qualtrics
}
\description{
A dataframe containing the state names (`state_label`) that correspond to the numeric state values (`state_value`) used by Qualtrics.
}
\keyword{datasets}
