% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs_tex.R
\name{subs_tex}
\alias{subs_tex}
\title{subs_tex}
\usage{
subs_tex(x)
}
\arguments{
\item{x}{a character scalar or vector to be formatted.}
}
\description{
A function for easily formatting a string as a subscript in LaTeX.
}
