% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supers_tex.R
\name{supers_tex}
\alias{supers_tex}
\title{supers_tex}
\usage{
supers_tex(x)
}
\arguments{
\item{x}{a character scalar or vector to be formatted.}
}
\description{
A function for easily formatting a string as a superscript in LaTeX format.
}
