% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_format.R
\name{text_format}
\alias{text_format}
\alias{bold}
\title{text_format}
\usage{
text_format(x, format = "italic", latex = FALSE)

bold(x)
}
\arguments{
\item{x}{a character scalar or vector to be formatted.}

\item{format}{a string specifying whether the text should be boled or italicized.}

\item{latex}{if TRUE, text is formatted using LaTeX formatting.}
}
\description{
A function for easily formatting strings as bold or italicized using markdown or LaTeX syntax. `bold(x)` is a shortcut for `text_format(x, format = "bold", latex = FALSE)` and `italic(x)` is a shortcut for `text_format(x, format = "italic", latex = FALSE)`. `bold_tex(x)` is a shortcut for `text_format(x, format = "bold", latex = TRUE)` and `italic_tex(x)` is a shortcut for `text_format(x, format = "italic", latex = TRUE)`.
}
