% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_std.R
\name{tidy_std}
\alias{tidy_std}
\title{tidy_std}
\usage{
tidy_std(model, conf.level = 0.95)
}
\arguments{
\item{model}{a lm or lmer model}

\item{conf.level}{the confidence level of the confidence intervals. Defaults to .95.}
}
\description{
Tidies lm and lmer model and includes the standardized coefficients, standard error, and confidence intervals..
}
