% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_alpha.R
\name{column_alpha}
\alias{column_alpha}
\alias{column_rij}
\alias{column_all}
\alias{column_o_h}
\alias{column_o_t}
\title{column_alpha}
\usage{
column_alpha(
  pattern,
  data,
  full = FALSE,
  verbose = FALSE,
  message = TRUE,
  na.rm = TRUE,
  return = "alpha",
  ci = FALSE,
  spround = FALSE
)

column_rij(pattern, data, ...)

column_all(pattern, data, ...)

column_o_h(pattern, data, ...)

column_o_t(pattern, data, ...)
}
\arguments{
\item{pattern}{a character string that is used to find the columns of interest. It can be a regular expression.}

\item{data}{a data frame.}

\item{full}{if TRUE, the full results of the reliability analysis produced by the `psych` package is returned. If FALSE, only the raw alpha value is returned.}

\item{verbose}{specifies whether all column names used should be listed in the message, regardless of length.}

\item{message}{if TRUE, messages are generated telling the user which columns were used to calculate Cronbach's Alpha.}

\item{na.rm}{a logical value indicating whether `NA` values should be removed prior to computation.}

\item{return}{a string indicating whether column_alpha should return Cronbach's alpha (`"alpha"`), the average correlation between the items (`"rij"`), omega hierarchical (`"o_h"`), omega total (`"o_t"`), or all four (`"all"`). Defaults to `"alpha"`.}

\item{ci}{a logical value indicating whether Cronbach's alpha should be returned with Feldt et al. (1987) confidence intervals. Defaults to `FALSE`.}

\item{spround}{a logical value indicating whether values should be rounded for printing. Defaults to FALSE.}

\item{...}{additional arguments passed to `column_alpha`.}
}
\description{
Estimates Cronbach's Alpha--an indicator of internal consistency--using only columns that have names that match a pattern. The analysis relies on `psych::alpha`.
}
