% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_rsq.R
\name{delta_rsq}
\alias{delta_rsq}
\title{delta_rsq}
\usage{
delta_rsq(models)
}
\arguments{
\item{models}{a list of linear models. Differences will be calculated in the order models are listed.}
}
\description{
Calculates change to R-Squared across two or more models.
}
