% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply-package.R
\docType{package}
\name{panoply-package}
\alias{panoply}
\alias{panoply-package}
\title{panoply: A Panoply of Miscellaneous Functions}
\description{
Contains a number of miscellaneous functions for making data analysis more efficient.
}
\author{
\strong{Maintainer}: Cameron Kay \email{cameronstuartkay@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
