% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_paren.R
\name{paste_paren}
\alias{paste_paren}
\title{paste_paren}
\usage{
paste_paren(x, y)
}
\arguments{
\item{x}{a scalar or an atomic vector.}

\item{y}{a second scalar or atomic vector that will be wrapped in parentheses.}
}
\description{
Combines two numbers (e.g., 10.12 and 2.22) by wrapping the latter in parentheses (e.g., 10.12(2.22)). This function was made to streamline the creation of tables that are formatted mean(sd).
}
