% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spround.R
\name{spround}
\alias{spround}
\title{spround}
\usage{
spround(x, digits = 2, leading0 = TRUE, less_than = FALSE)
}
\arguments{
\item{x}{a numeric/character scalar or vector.}

\item{digits}{a numeric scalar indicating the number of decimal places to round to.}

\item{leading0}{a logical scalar indicating whether leading zeros should be included.}

\item{less_than}{a logical scalar indicating whether rounded values equal to zero should be represented as less than a certain value (e.g., 0.000 becomes < .001).}
}
\description{
Round and specify knitted decimal places all in one step.
}
\examples{
spround(x = c(0.1078, 9.9930, 42.4242), digits = 2, leading0 = TRUE)
spround(x = c(0.1078, 9.9930, 42.4242), digits = 2, leading0 = FALSE)
}
