% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark.R, R/text_format.R
\name{dark_triad}
\alias{dark_triad}
\alias{dark_tetrad}
\alias{text_format}
\alias{bold}
\alias{bold_tex}
\alias{italic}
\alias{italic_tex}
\alias{supers}
\alias{supers_tex}
\alias{subs}
\alias{subs_tex}
\title{text_format}
\usage{
dark_triad(x, ...)

dark_tetrad(x, ...)

text_format(x, format = "italic", latex = FALSE)

bold(x)

bold_tex(x)

italic(x)

italic_tex(x)

supers(x)

supers_tex(x)

subs(x)

subs_tex(x)
}
\arguments{
\item{x}{a character scalar or vector to be formatted.}

\item{format}{a string specifying whether the text should be boled or italicized.}

\item{latex}{if TRUE, text is formatted using LaTeX formatting.}
}
\description{
A function for easily formatting strings as bold or italicized using markdown or LaTeX syntax. `bold(x)` is a shortcut for `text_format(x, format = "bold", latex = FALSE)` and `italic(x)` is a shortcut for `text_format(x, format = "italic", latex = FALSE)`. `bold_tex(x)` is a shortcut for `text_format(x, format = "bold", latex = TRUE)` and `italic_tex(x)` is a shortcut for `text_format(x, format = "italic", latex = TRUE)`.
}
