% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zo.R
\name{zo}
\alias{zo}
\title{zo}
\usage{
zo(data, cols, split, adjust.p = "none", spround = TRUE, pasterisk = TRUE, ...)
}
\arguments{
\item{data}{a data frame.}

\item{cols}{the names of the columns to be included in the correlation matrix. Defaults to every column except the split column (if a split column is provided).}

\item{split}{an optional string indicating the column that includes the grouping variable for producing different matrices above and below the diagonal.}

\item{adjust.p}{a string indicating what type of correction for multiple comparisons should be used. Defaults to "none."}

\item{spround}{a logical value indicating whether values should be rounded for printing. Defaults to TRUE.}

\item{pasterisk}{a logical value indicating whether p-values should be replaced with asterisks and combined with the regression coefficients. Defaults to TRUE.}

\item{...}{optional arguments to be passed to pasterisk.}
}
\description{
Creates a zero-order correlation matrix.
}
